refactor: move nationalities slice into core application
