+++ backend/src/api/dependencies/use_cases.py
@@ -2,6 +2,7 @@ from auth.application.use_cases.login import LoginAdminUseCase, LoginUserUseCase
 from auth.infrastructure.repositories.sqlalchemy_auth_account_repository import (
     SqlAlchemyAuthAccountRepository,
 )
+from core.application.use_cases.get_nationalities_usecase import GetNationalitiesUseCase
 from core.application.use_cases.get_penas_usecase import GetPenasUseCase
 from core.application.use_cases.get_season_match_insights_usecase import (
     GetSeasonMatchInsightsUseCase,
@@ -10,7 +11,6 @@ from fastapi import Depends
 from persistence.application.use_cases.generate_pena_link_token_usecase import (
     GeneratePenaLinkTokenUseCase,
 )
-from persistence.application.use_cases.get_nationalities_usecase import GetNationalitiesUseCase
 from persistence.application.use_cases.get_pena_players_usecase import GetPenaPlayersUseCase
 from persistence.application.use_cases.get_player_profile_usecase import (
     GetPlayerProfileUseCase,
