+++ backend/src/api/interface/controller/v1/catalogs_controller.py
@@ -1,6 +1,6 @@
 from api.dependencies.use_cases import get_nationalities_use_case
+from core.application.use_cases.get_nationalities_usecase import GetNationalitiesUseCase
 from fastapi import APIRouter, Depends
-from persistence.application.use_cases.get_nationalities_usecase import GetNationalitiesUseCase
 
 router = APIRouter()
