+++ backend/src/core/application/ports/__init__.py
@@ -1,4 +1,5 @@
+from .nationality_query_port import NationalityQueryPort
 from .pena_query_port import PenaQueryPort
 from .season_match_insights_port import SeasonMatchInsightsPort
 
-__all__ = ["PenaQueryPort", "SeasonMatchInsightsPort"]
+__all__ = ["NationalityQueryPort", "PenaQueryPort", "SeasonMatchInsightsPort"]
