+++ backend/src/core/application/ports/nationality_query_port.py
@@ -1,0 +1,5 @@
+from typing import Protocol
+
+
+class NationalityQueryPort(Protocol):
+    def list_names(self) -> list[str]: ...
