+++ backend/src/core/application/use_cases/get_nationalities_usecase.py
@@ -1,0 +1,9 @@
+from core.application.ports.nationality_query_port import NationalityQueryPort
+
+
+class GetNationalitiesUseCase:
+    def __init__(self, repository: NationalityQueryPort):
+        self.repository = repository
+
+    def execute(self) -> list[str]:
+        return self.repository.list_names()
