+++ backend/src/persistence/infrastructure/repository/db/nationality_query_repository.py
@@ -1,6 +1,4 @@
-from persistence.application.ports.nationality_query_port import (
-    NationalityQueryPort,
-)
+from core.application.ports.nationality_query_port import NationalityQueryPort
 from persistence.domain.entity import Nationality
 from sqlalchemy import select
 from sqlalchemy.orm import Session
