+++ backend/tests/test_get_nationalities_use_case.py
@@ -1,4 +1,4 @@
-from persistence.application.use_cases.get_nationalities_usecase import GetNationalitiesUseCase
+from core.application.use_cases.get_nationalities_usecase import GetNationalitiesUseCase
 
 
 class _Repo:
